feat: add notification, rules and devices popups
