+++ frontend/src/scripts/presentation/notification/ComposeNotification.ts
@@ -1,0 +1,48 @@
+import RequestHelper, { alarmHost, alarmPort } from "@/utils/RequestHelper";
+import { DeviceTypeConverter, MeasureConverter, ObjectClassConverter } from "domain/dist/utils";
+import { DeviceType } from "domain/dist/domain/device/core";
+import type { Exceeding, Intrusion } from "domain/dist/domain/anomaly/core";
+import type { NotificationFactory } from "domain/dist/domain/alarm-system/factories";
+import { NotificationFactoryImpl } from "domain/dist/domain/alarm-system/factories";
+import type { AnomalyFactory } from "domain/dist/domain/anomaly/factories";
+import { AnomalyFactoryImpl } from "domain/dist/domain/anomaly/factories";
+import type { DeviceIdFactory } from "domain/dist/domain/device/factories";
+import { DeviceIdFactoryImpl } from "domain/dist/domain/device/factories";
+
+const notificationFactory: NotificationFactory = new NotificationFactoryImpl()
+const anomalyFactory: AnomalyFactory = new AnomalyFactoryImpl()
+const deviceIdFactory: DeviceIdFactory = new DeviceIdFactoryImpl()
+
+export async function composeNotification(notification: any) {
+  await RequestHelper.get(`http://${alarmHost}:${alarmPort}/anomalies/` + notification.anomalyId)
+    .then((anomaly: any) => {
+      switch (DeviceTypeConverter.convertToDeviceType(anomaly.data.deviceId.type)) {
+        case DeviceType.CAMERA:
+          return notificationFactory.createIntrusionNotification(notification._id, composeIntrusion(anomaly.data))
+        case DeviceType.SENSOR:
+          return notificationFactory.createExceedingNotification(notification._id, composeExceeding(anomaly.data))
+      }
+    })
+    .catch(error => {
+      console.log(error)
+    })
+}
+
+function composeIntrusion(intrusion: any): Intrusion {
+  return anomalyFactory.createIntrusion(
+    intrusion._id,
+    deviceIdFactory.createCameraId(intrusion.deviceId.code),
+    new Date(intrusion.timestamp),
+    ObjectClassConverter.convertToObjectClass(intrusion.intrusionObject)
+  )
+}
+
+function composeExceeding(exceeding: any): Exceeding {
+  return anomalyFactory.createExceeding(
+    exceeding._id,
+    deviceIdFactory.createSensorId(exceeding.deviceId.code),
+    new Date(exceeding.timestamp),
+    MeasureConverter.convertToMeasure(exceeding.measure),
+    exceeding.value
+  )
+}
