+++ frontend/src/views/NotificationView.vue
@@ -1,74 +1,23 @@
 <script setup lang="ts">
 import { onMounted, ref } from 'vue'
-
-import type { Exceeding, Intrusion } from '@domain/anomaly/core'
 import type { Notification } from '@domain/alarm-system/core'
-import { DeviceType } from '@domain/device/core'
-import { type AnomalyFactory, AnomalyFactoryImpl } from '@domain/anomaly/factories'
-import { type DeviceIdFactory, DeviceIdFactoryImpl } from '@domain/device/factories'
-import { type NotificationFactory, NotificationFactoryImpl } from '@domain/alarm-system/factories'
-import { ObjectClassConverter } from '@utils/ObjectClassConverter.js'
-import { DeviceTypeConverter } from '@utils/DeviceTypeConverter.js'
-import { MeasureConverter } from '@utils/MeasureConverter.js'
 import RequestHelper, { alarmHost, alarmPort } from '@/utils/RequestHelper'
 import NotificationBadge from '@/components/notification/NotificationBadge.vue'
-
-const notificationFactory: NotificationFactory = new NotificationFactoryImpl()
-const anomalyFactory: AnomalyFactory = new AnomalyFactoryImpl()
-const deviceIdFactory: DeviceIdFactory = new DeviceIdFactoryImpl()
+import { composeNotification } from '@/scripts/presentation/notification/ComposeNotification'
 
 let notifications: ref<Notification[]> = ref<Notification[]>([])
 
 async function getNotifications() {
   await RequestHelper.get(`http://${alarmHost}:${alarmPort}/notifications`)
     .then((res: any) => {
+      notifications.value = []
       for (let i = 0; i < res.data.length; i++) {
-        composeNotification(res.data[i])
+        notifications.value.push(composeNotification(res.data[i]))
       }
     })
     .catch(error => {
       console.log(error)
     })
-}
-
-async function composeNotification(notification: any) {
-  await RequestHelper.get(`http://${alarmHost}:${alarmPort}/anomalies/` + notification.anomalyId)
-    .then((anomaly: any) => {
-      switch (DeviceTypeConverter.convertToDeviceType(anomaly.data.deviceId.type)) {
-        case DeviceType.CAMERA:
-          notifications.value.push(
-            notificationFactory.createIntrusionNotification(notification._id, composeIntrusion(anomaly.data))
-          )
-          break
-        case DeviceType.SENSOR:
-          notifications.value.push(
-            notificationFactory.createExceedingNotification(notification._id, composeExceeding(anomaly.data))
-          )
-          break
-      }
-    })
-    .catch(error => {
-      console.log(error)
-    })
-}
-
-function composeIntrusion(intrusion: any): Intrusion {
-  return anomalyFactory.createIntrusion(
-    intrusion._id,
-    deviceIdFactory.createCameraId(intrusion.deviceId.code),
-    new Date(intrusion.timestamp),
-    ObjectClassConverter.convertToObjectClass(intrusion.intrusionObject)
-  )
-}
-
-function composeExceeding(exceeding: any): Exceeding {
-  return anomalyFactory.createExceeding(
-    exceeding._id,
-    deviceIdFactory.createSensorId(exceeding.deviceId.code),
-    new Date(exceeding.timestamp),
-    MeasureConverter.convertToMeasure(exceeding.measure),
-    exceeding.value
-  )
 }
 
 onMounted(async () => {
